+++ CHANGELOG.md
@@ -13,7 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 > - **Fixed**: for any bug fixes.
 > - **Security**: in case of vulnerabilities.
 
-## [3.3.0] - 
+## [3.x.x] - 
+
+- Adds metrics. If run with the `--export :2112` flag (`:2112` is a port),
+  metrics are exposed as Prometheus metrics on that port and can be HTTP
+  GET-ted.  
+
+
+## [3.3.0] - 2020-02-17
 
 - Added: Logs are now rotated. Settings are currently hard-coded at 4 files of 5MB
   each, so logs shouldn't take up more than 20MB.  I'm going to see how many